Convert PolarAngleAxisSpec.js -> PolarAngleAxis.spec.tsx
